SponsorDonation page with placeholder form and buttons
