--- bravechild/src/components/BoxContainer.js
@@ -1,21 +1,0 @@
-import React from 'react'
-import "./styles/BoxContainer.css"
-
-const BoxContainer = () => {
-    const boxes = [1, 2, 3]
-
-    return (
-        <div className='box-container'>
-            {boxes.map((box, index) => (
-                <div key={index} className='box'>
-                    <p>Box {box}</p>
-                    <div className='placeholder'>
-                        <p>Image/Payment Button</p>
-                    </div>
-                </div>
-            ))}
-        </div>
-    );
-};
-
-export default BoxContainer;
