+++ bravechild/src/styles/DonationForm.css
@@ -1,4 +1,3 @@
-
 .form-container {
     max-width: 500px;
     margin: 0 auto;
@@ -52,7 +51,7 @@
     font-size: 1rem;
     font-weight: bold;
     color: #fff;
-    background-color: #007BFF;
+    background-color: #4CAF50;
     border: none;
     border-radius: 4px;
     cursor: pointer;
@@ -60,5 +59,23 @@
 }
 
 .form-button:hover {
-    background-color: #0056b3;
+    background-color: #45a049;
+}
+
+.payment-options {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+    margin-top: 20px;
+}
+
+.payment-image {
+    width: 100px;
+    height: 100px;
+    object-fit: contain;
+    cursor: pointer;
+    border: 1px solid #000;
+    border-radius: 15%;
+    padding: 5px;
+    background-color:#fff; 
 }
