+++ bravechild/src/styles/SponsorDonations.css
@@ -1,0 +1,3 @@
+.title {
+    text-align: center;
+}
